Add firmware store
Arduino projects support auto-upgrade of firmware. The protocol is well
defined. There is an existent implementation on client-side for arduino
devices, however implementations on server side are designed to be run
in the IDE or just snippets (in languages like PHP). The firmware store
in this PR is a way to:
1) Protect the firmware so that only authorized devices can access it
2) Provide version discovery feature (conform to existent standard)