Revert "Keep last request per device even if expired"
Context: reboot of device, load of last published report to
reinitialization of settings as per when last running.

Now its the responsibility of the devices to publish a full report
regularly, containing all the properties of all the actors. Given that a
report can be partial (i.e. contain not all the actors properties,
keeping the last does not guarantee a good report being kept).