Fix ServerSpec
The reason why the test was unstable is timing.
The cleanup takes place regularly every C secs (0.5 secs). The retention
is of R secs (2 secs).

A. If C (0.1) secs is far from R (2) secs, it makes sense to check in
between periods of R secs (because many cleanups will take place in
between).

B. If C (2) secs is close to R (2) secs, it's of no point to try to
check cleanups performed in between R secs (a few or no cleanups will
take place in between).

C. Then, if C is somewhere ~0.5, (as before the fix) it is very
important the synchronization of cleanups triggered and retention period
reached. So this was making the test unstable.

This change puts the values in situation A.